Fix anonymization
